+++ src/agents/__tests__/IssueReporter.security.test.ts
@@ -4,29 +4,25 @@
  * Validates that attachScreenshot does NOT upload screenshot data to GitHub Gists.
  * 'Secret' gists have publicly accessible URLs, making them unsuitable for
  * sensitive screenshot data that may contain credentials, PII, or internal state.
+ *
+ * Test approach: attachScreenshot is a public method on IssueReporter so no
+ * private-member casting is required. The gists.create spy is verified to be
+ * unused by injecting the mock Octokit at module level via jest.mock('@octokit/rest').
+ * This replaces the earlier (reporter as any).octokit assignment which was brittle —
+ * "octokit" is a private field name and an implementation detail.
  */
 
 import { IssueReporter } from '../IssueReporter';
 import { GitHubConfig } from '../../models/Config';
 
-// Minimal GitHub config
-function makeConfig(): GitHubConfig {
+// Mock @octokit/rest so no real network calls are made and so we can assert
+// that gists.create is never invoked. The spy is defined at module scope so
+// it persists across the jest.mock() hoisting boundary.
+const gistsCreateSpy = jest.fn();
+
+jest.mock('@octokit/rest', () => {
   return {
-    token: 'test-token',
-    owner: 'test-owner',
-    repo: 'test-repo',
-  } as GitHubConfig;
-}
-
-describe('IssueReporter - screenshot security (issue #98)', () => {
-  let reporter: IssueReporter;
-  let mockOctokit: Record<string, unknown>;
-
-  beforeEach(() => {
-    // Spy to ensure gists.create is NEVER called
-    const gistsCreateSpy = jest.fn();
-
-    mockOctokit = {
+    Octokit: jest.fn().mockImplementation(() => ({
       rest: {
         gists: {
           create: gistsCreateSpy,
@@ -42,21 +38,31 @@ describe('IssueReporter - screenshot security (issue #98)', () => {
           }),
         },
       },
-    };
+    })),
+  };
+});
 
+// Minimal GitHub config
+function makeConfig(): GitHubConfig {
+  return {
+    token: 'test-token',
+    owner: 'test-owner',
+    repo: 'test-repo',
+  } as GitHubConfig;
+}
+
+describe('IssueReporter - screenshot security (issue #98)', () => {
+  let reporter: IssueReporter;
+
+  beforeEach(() => {
+    gistsCreateSpy.mockClear();
     reporter = new IssueReporter(makeConfig());
-    // Replace internal octokit with our spy
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (reporter as any).octokit = mockOctokit;
   });
 
   describe('attachScreenshot', () => {
     it('should NOT call gists.create (no Gist upload)', async () => {
-      const gistsCreateSpy = (mockOctokit.rest as any).gists.create as jest.Mock;
-
-      // Call attachScreenshot with a local path
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      const result = await (reporter as any).attachScreenshot(42, '/tmp/screenshot.png');
+      // attachScreenshot is a public method — no (as any) cast needed
+      const result = await reporter.attachScreenshot(42, '/tmp/screenshot.png');
 
       expect(gistsCreateSpy).not.toHaveBeenCalled();
       // Should return the local path (or undefined) - never a gist URL
@@ -69,18 +75,16 @@ describe('IssueReporter - screenshot security (issue #98)', () => {
     it('should return the local screenshot path (not a remote URL)', async () => {
       const screenshotPath = '/tmp/test-screenshot.png';
 
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      const result = await (reporter as any).attachScreenshot(42, screenshotPath);
+      // attachScreenshot is a public method — no (as any) cast needed
+      const result = await reporter.attachScreenshot(42, screenshotPath);
 
       // Result must either be the local path or undefined - never a Gist URL
       expect(result === screenshotPath || result === undefined).toBe(true);
     });
 
     it('should not upload binary data to any external service', async () => {
-      const gistsCreateSpy = (mockOctokit.rest as any).gists.create as jest.Mock;
-
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      await (reporter as any).attachScreenshot(99, '/tmp/sensitive-screenshot.png');
+      // attachScreenshot is a public method — no (as any) cast needed
+      await reporter.attachScreenshot(99, '/tmp/sensitive-screenshot.png');
 
       // Primary assertion: Gist upload must not occur
       expect(gistsCreateSpy).not.toHaveBeenCalled();
